filter unliking/deleting for like/posts/reposts/all
currently the app works on all types of skeets - posts, reposts, and
likes.  The new -t / --post-type allows you to just unlike likes, or
just remove reposts...

This is helpful for removing reposts if you normally use the -p option
to keep popular stuff... as that popular filter runs against the like
count of the original post.